// backend/test/setup.d.ts
export declare const prismaMock: {
    user: {
        findUnique: import("vitest").Mock<import("@vitest/spy").Procedure>;
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
        count: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    story: {
        findUnique: import("vitest").Mock<import("@vitest/spy").Procedure>;
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
        count: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    transport: {
        findUnique: import("vitest").Mock<import("@vitest/spy").Procedure>;
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    message: {
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    event: {
        findUnique: import("vitest").Mock<import("@vitest/spy").Procedure>;
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    booking: {
        findUnique: import("vitest").Mock<import("@vitest/spy").Procedure>;
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    community: {
        findUnique: import("vitest").Mock<import("@vitest/spy").Procedure>;
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    like: {
        findUnique: import("vitest").Mock<import("@vitest/spy").Procedure>;
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    comment: {
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
        delete: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    badge: {
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    userBadge: {
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    notification: {
        findMany: import("vitest").Mock<import("@vitest/spy").Procedure>;
        create: import("vitest").Mock<import("@vitest/spy").Procedure>;
        update: import("vitest").Mock<import("@vitest/spy").Procedure>;
    };
    $connect: import("vitest").Mock<import("@vitest/spy").Procedure>;
    $disconnect: import("vitest").Mock<import("@vitest/spy").Procedure>;
    $transaction: import("vitest").Mock<import("@vitest/spy").Procedure>;
};